tree: extract `data_type` module from cass_types
Most of the `cass_types` module is CassDataType-related entities.
Those were moved to the newly created `cql_types::data_type` module.
