+++ scylla-rust-wrapper/src/cql_types/tuple.rs
@@ -1,7 +1,6 @@
 use crate::argconv::*;
 use crate::cass_error::CassError;
-use crate::cass_types::CassDataType;
-use crate::cass_types::CassDataTypeInner;
+use crate::cql_types::data_type::{CassDataType, CassDataTypeInner};
 use crate::cql_types::value;
 use crate::cql_types::value::CassCqlValue;
 use crate::types::*;
@@ -136,8 +135,11 @@ make_binders!(user_type, cass_tuple_set_user_type);
 
 #[cfg(test)]
 mod tests {
-    use crate::cass_types::{
-        CassValueType, cass_data_type_add_sub_type, cass_data_type_free, cass_data_type_new,
+    use crate::{
+        cass_types::CassValueType,
+        cql_types::data_type::{
+            cass_data_type_add_sub_type, cass_data_type_free, cass_data_type_new,
+        },
     };
 
     use super::{cass_tuple_data_type, cass_tuple_new};
