+++ scylla-rust-wrapper/src/cql_types/value.rs
@@ -11,7 +11,8 @@ use scylla::serialize::writers::{CellWriter, WrittenCellProof};
 use scylla::value::{CqlDate, CqlDecimal, CqlDuration};
 use uuid::Uuid;
 
-use crate::cass_types::{CassDataType, CassValueType};
+use crate::cass_types::CassValueType;
+use crate::cql_types::data_type::CassDataType;
 
 /// A narrower version of rust driver's CqlValue.
 ///
@@ -416,8 +417,11 @@ mod tests {
     use scylla::value::{CqlDate, CqlDecimal, CqlDuration};
 
     use crate::{
-        cass_types::{CassDataType, CassDataTypeInner, CassValueType, MapDataType, UdtDataType},
-        cql_types::value::{CassCqlValue, is_type_compatible},
+        cass_types::CassValueType,
+        cql_types::{
+            data_type::{CassDataType, CassDataTypeInner, MapDataType, UdtDataType},
+            value::{CassCqlValue, is_type_compatible},
+        },
     };
 
     fn all_value_data_types() -> Vec<CassDataType> {
