+++ scylla-rust-wrapper/src/prepared.rs
@@ -4,7 +4,7 @@ use std::{os::raw::c_char, sync::Arc};
 use crate::{
     argconv::*,
     cass_error::CassError,
-    cass_types::{CassDataType, get_column_type},
+    cql_types::data_type::{CassDataType, get_column_type},
     query_result::CassResultMetadata,
     statement::{BoundPreparedStatement, CassStatement},
     types::size_t,
