+++ scylla-rust-wrapper/src/session.rs
@@ -2,8 +2,8 @@ use crate::argconv::*;
 use crate::batch::CassBatch;
 use crate::cass_error::*;
 use crate::cass_metrics_types::CassMetrics;
-use crate::cass_types::get_column_type;
 use crate::cluster::CassCluster;
+use crate::cql_types::data_type::get_column_type;
 use crate::cql_types::uuid::CassUuid;
 use crate::exec_profile::{CassExecProfile, ExecProfileName, PerStatementExecProfile};
 use crate::future::{CassFuture, CassFutureResult, CassResultValue};
